% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot.R
\name{render.heatmap}
\alias{render.heatmap}
\title{Render Heatmap Covid plot}
\usage{
render.heatmap(path.data, sheet.name, min.depth = 100, show = "ac")
}
\arguments{
\item{path.data}{The path of the excel file to use}

\item{sheet.name}{The sheet nome of the excel to process}

\item{min.depth}{The minimum depth, positions with less depth than the threshold will be displayed in white on the plot.}

\item{show}{This option allows you to select the items to be displayed (acide nucleic (ac) or amino acid (aa)), by default the function show ac information. You can choose between 'ac' and 'aa'}
}
\value{
A beautiful plot
}
\description{
Create a Covid heatmap for show SNP evolution in genome
}
\examples{
render.heatmap(path.data='/path/to/excel', sheet.name='final',  min.depth=100)
}
