% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot.R
\name{render.plot}
\alias{render.plot}
\title{Render Covid plot}
\usage{
render.plot(
  path.data,
  sheet.name,
  min.frequency = 0,
  min.depth = 100,
  show = "all"
)
}
\arguments{
\item{path.data}{The path of the excel file to use}

\item{sheet.name}{The sheet nome of the excel to process}

\item{min.frequency}{The min frequency (SNPs with a frequency below the threshold will not be displayed)}

\item{min.depth}{The minimum depth, positions with less depth than the threshold will be displayed in white on the plot.}

\item{show}{This option allows you to select the items to be displayed, by default the function show all genome. You can choose between 'ORF1ab', 'S', 'ORF3a', 'E', 'M', 'ORF6', 'ORF7a', 'ORF8', 'N', 'ORF10'}
}
\value{
A beautiful plot
}
\description{
Create a Covid plot for show SNP evolution in genome
}
\examples{
render.plot(path.data='/path/to/excel', sheet.name='final', min.frequency=10, min.depth=100)
}
